--- src/main/java/racingcargame/model/CarGameRaceCount.java
@@ -1,21 +1,0 @@
-package racingcargame.model;
-
-public class CarGameRaceCount implements RaceCount{
-    private static final int GAME_OVER = 0;
-
-    private int count;
-
-    public CarGameRaceCount(int count) {
-        this.count = count;
-    }
-
-    @Override
-    public void reduceRaceCount() {
-        count--;
-    }
-
-    @Override
-    public boolean isZeroRaceCount() {
-        return count == GAME_OVER;
-    }
-}
